+++ src/main/java/com/grinder/service/ImageService.java
@@ -2,39 +2,16 @@ package com.grinder.service;
 
 import com.grinder.domain.entity.Image;
 import com.grinder.domain.enums.ContentType;
-import com.grinder.repository.ImageRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
 
 import java.util.List;
 
-@Service
-@RequiredArgsConstructor
-public class ImageService {
-    ImageRepository imageRepository;
-
-    public Image findImage(String imageId) {
-        return imageRepository.findById(imageId).orElseThrow(() -> new IllegalArgumentException("image id(" + imageId + ")를 찾울 수 없습니다."));
-    }
-
-    public List<Image> findAllImage(String contentId, ContentType contentType) {
-        return imageRepository.findByContentIdAndContentType(contentId, contentType);
-    }
-
-    public void saveFeedImage(String contentId, ContentType contentType, List<String> imageUrlList) {
-        for (String imageUrl : imageUrlList) {
-            imageRepository.save(
-                    Image.builder()
-                            .imageUrl(imageUrl)
-                            .contentId(contentId)
-                            .contentType(contentType)
-                            .build()
-                    );
-        }
-    }
-
-    public void deleteFeedImage(String contentId, ContentType contentType) {
-        List<Image> ImageList = findAllImage(contentId, contentType);
-        imageRepository.deleteAll(ImageList);
-    }
+public interface ImageService {
+    // imageId로 이미지 찾기
+    Image findImage(String imageId);
+    // 해당 컨텐츠(피드)에서 사용한 모든 이미지 찾기
+    List<Image> findAllImage(String contentId, ContentType contentType);
+    // 해당 컨텐츠(피드)에서 사용할 이미지 모두 저장
+    void saveFeedImage(String contentId, ContentType contentType, List<String> imageUrlList);
+    // 해당 컨텐츠(피드)에서 사용한 이미지 모두 삭제
+    void deleteFeedImage(String contentId, ContentType contentType);
 }
